// src/app/home/models/tb-buyer.ts
export class TbBuyer {
  byrSerial?: number= null;
  byrAddr1?: string= null;
  byrAmphur?: string= null;
  byrCode?: string= null;
  byrFax?: string= null;
  byrMobile?: string= null;
  byrNameEng?: string= null;
  byrNameTh?: string= null;
  byrPhone?: string= null;
  byrPostcode?: string= null;
  byrPplno?: string= null;
  byrProvinceId?: string= null;
  byrTumbon?: string= null;
  byrbranchCode?: string= null;
  byrbranchName?: string= null;
  byrbusinessId?: number= null;
  byrcountryId?: number= null;
  byrfzId?: number= null;
  byrplantId?: number= null;
  byrstsId?: number= null;
  byrtaxId?: string= null;
  byrtaxIncentive?: string= null;
  dayStart?: number= null;
  orderDays?: number= null;
  orderForm?: string= null;
  createBy?: number= null;
  updateBy?: number= null;
}